## Refactor OpenAI Dependencies
- **New Modules:** Created `openai_chat.py` and `openai_threads.py` to streamline interactions with OpenAI's chat and threads APIs respectively.
- **Dependency Updates:**
    - Refactored `chat.py` and `threads.py` to import from the new modules.
    - Removed chat and threads specific dependencies from `openai_clients.py`.
